some small changes to monitoring and validation scripts, plus configs for latest rerecos
Conflicts:
	ZFitter/script/makeTable.sh
	ZFitter/src/ZFit_class.cc

+++ ZFitter/script/validation.sh
@@ -9,9 +9,10 @@ FLOATTAILSTEXT="Fixed Tail"
 runRangesFile=data/runRanges/monitoring.dat 
 baseDir=test
 xVar=runNumber
-pdateOnly="--updateOnly"
+updateOnly="--updateOnly"
 GITTAG="lcorpe:topic-quickrereco-lcsh-fix-rebase" #should eventually get this automatically from file or option
 GLOBALTAG="74X-lcdataRun2-lcPrompt-lcv0" #should eventually get this automatically from file or option
+imgFormat="eps"
 
 # VALIDATION=y
 # STABILITY=y
@@ -29,10 +30,13 @@ usage(){
     echo " --validation        "
     echo " --stability         "
     echo " --slides            "
+    echo " --ref compare to a reference rereco            "
     echo " --noPU            "
     echo " --floatTails            "
     echo " --baseDir arg (=${baseDir})  base directory for outputs"
     echo " --rereco arg  name of the tag used fro the rereco"
+    echo " --imgFormat arg eps or pdf"
+		echo " --dryRun make all the directories but don't run fits"
 }
 desc(){
     echo "#####################################################################"
@@ -45,7 +49,7 @@ desc(){
 
 
 # options may be followed by one colon to indicate they have a required argument
-if ! options=$(getopt -u -o hf: -l help,runRangesFile:,selection:,invMass_var:,puName:,dataName:,baseDir:,rereco:,validation,stability,slides,noPU,floatTails -- "$@")
+if ! options=$(getopt -u -o hf: -l help,runRangesFile:,selection:,invMass_var:,puName:,dataName:,baseDir:,rereco:,validation,stability,slides,noPU,floatTails,dryRun,imgFormat:,ref: -- "$@")
 then
     # something went wrong, getopt will put out an error message for us
     exit 1
@@ -60,15 +64,19 @@ do
         -f) configFile=$2; shift;;
         --invMass_var) invMass_var=$2; echo "[OPTION] invMass_var = ${invMass_var}"; shift;;
 	--puName) puName=$2; shift;;
+	--ref) REF=$2; shift;;
 	--dataName) dataName=$2; shift;;
 	--runRangesFile) runRangesFile=$2; echo "[OPTION] runRangesFile = ${runRangesFile}"; shift;;
 	--baseDir) baseDir=$2; echo "[OPTION] baseDir = $baseDir"; shift;;
+	--imgFormat) imgFormat=$2; echo "[OPTION] imgFormat = $imgFormat"; shift;;
 	--rereco) rereco=$2; echo "[OPTION] rereco = $rereco"; shift;;
 	--validation) VALIDATION=y;;
 	--stability)  STABILITY=y;;
 	--slides)     SLIDES=y;;
 	--noPU)     NOPUOPT="--noPU"; NOPU="noPU";;
 	--floatTails)    FLOATTAILSOPT="--fit_type_value=0"; FLOATTAILS="floatingTail"; FLOATTAILSTEXT="Floating Tail";;
+	--dryRun)     DRYRUN="1";;
+	--selection) selection=$2; echo "test selection" $selection ; shift;;
         (--) shift; break;;
         (-*) usage; echo "$0: error - unrecognized option $1" 1>&2; usage >> /dev/stderr; exit 1;;
         (*) break;;
@@ -79,9 +87,9 @@ done
 if [ -n "$NOPU" ];then
   baseDir="$baseDir-$NOPU" 
 fi
-if [ -n "$FLOATTAILS" ];then
-  baseDir="$baseDir-$FLOATTAILS" 
-fi
+#if [ -n "$FLOATTAILS" ];then
+#  baseDir="$baseDir-$FLOATTAILS" 
+#fi
 
 
 if [ -z "${VALIDATION}" -a -z "${STABILITY}" -a -z "${SLIDES}" ];then
@@ -99,15 +107,15 @@ if [ -z "${configFile}" ];then
     exit 1
 fi
 
-case ${selection} in
-    WP80_PU)
-        ;;
-    WP90_PU)
-	;;
-    *)
-			exit 1
-        ;;
-esac
+#case ${selection} in
+#    WP80_PU)
+#        ;;
+#    WP90_PU)
+#	;;
+#    *)
+#			exit 1
+#        ;;
+#esac
 
 
 
@@ -163,6 +171,21 @@ if [ ! -e "${outDirData}/fitres" ];then mkdir ${outDirData}/fitres -p; fi
 if [ ! -e "${outDirData}/img" ];then mkdir ${outDirData}/img -p; fi
 if [ ! -e "${outDirData}/log" ];then mkdir ${outDirData}/log -p; fi
 
+#if [ -z "$FLOATTAILSOPT" ] ;then cp $REF/MC/DYJets_madgraph-RunIISpring15DR74-Asym25ns-allRange/pufile_tot/$selection/${invMass_var}/fitres/* ${outDirData}/fitres/. || exit 1 fi
+if [[ -n "$REF" ]] && [[ -z "$FLOATTAILSOPT" ]]; then
+refDir=$REF/dato/$REF/${selection}/${invMass_var}
+count=`ls -l $refDir/fitres | wc -l`
+echo $count
+if (( $count == 0 )); then
+ ls -A $refDir/fitres
+echo "[ERROR] specified Ref does not have anything in $refDir/fitres! exit!"
+exit 1
+fi
+outDirMC=${refDir}
+fi
+# cp $REF/MC/DYJets_madgraph-RunIISpring15DR74-Asym25ns-allRange/pufile_tot/$selection/${invMass_var}/fitres/* ${outDirMC}/fitres/. 
+if [ -n "$DRYRUN" ]; then exit 0; fi
+
 # keep track of the MC used to take the tail parameter for data
 echo "$outDirMC" > $outDirData/whichMC.txt
 
@@ -170,10 +193,14 @@ if [ -n "$VALIDATION" ];then
     mcSample=${mcName}
     dataSample=${dataName}
     regionFile=data/regions/validation.dat
+echo "
     ./bin/ZFitter.exe -f ${configFile} --regionsFile ${regionFile}  --invMass_var ${invMass_var} \
 	--outDirFitResMC=${outDirMC}/fitres --outDirFitResData=${outDirData}/fitres \
-	--outDirImgMC=${outDirMC}/img   --outDirImgData=${outDirData}/img  --commonCut $commonCut $NOPUOPT $FLOATTAILSOPT > ${outDirData}/log/validation.log||  exit 1
-
+	--outDirImgMC=${outDirMC}/img   --outDirImgData=${outDirData}/img  --commonCut $commonCut $NOPUOPT $FLOATTAILSOPT --imgFormat $imgFormat --selection $selection > ${outDirData}/log/validation.log||  exit 1
+"
+    ./bin/ZFitter.exe -f ${configFile} --regionsFile ${regionFile}  --invMass_var ${invMass_var} \
+	--outDirFitResMC=${outDirMC}/fitres --outDirFitResData=${outDirData}/fitres \
+	--outDirImgMC=${outDirMC}/img   --outDirImgData=${outDirData}/img  --commonCut $commonCut $NOPUOPT $FLOATTAILSOPT --imgFormat $imgFormat --selection $selection> ${outDirData}/log/validation.log||  exit 1
 
 
     ./script/makeTable.sh --regionsFile ${regionFile}  --commonCut=${commonCut} \
@@ -187,12 +214,23 @@ regionFile=data/regions/stability.dat
 ./bin/ZFitter.exe -f ${configFile} --regionsFile ${regionFile}  --runRangesFile ${runRangesFile} \
     $updateOnly --invMass_var ${invMass_var} \
     --outDirFitResMC=${outDirMC}/fitres --outDirFitResData=${outDirData}/fitres \
-    --outDirImgMC=${outDirMC}/img    --outDirImgData=${outDirData}/img --commonCut=${commonCut}> ${outDirData}/log/stability.log || exit 1
+    --outDirImgMC=${outDirMC}/img    --outDirImgData=${outDirData}/img --commonCut=${commonCut}  \
+		--imgFormat $imgFormat --selection $selection|| exit 1 #> ${outDirData}/log/stability.log || exit 1
+echo " 
+    ./bin/ZFitter.exe -f ${configFile} --regionsFile ${regionFile}  --invMass_var ${invMass_var} \
+	--outDirFitResMC=${outDirMC}/fitres --outDirFitResData=${outDirData}/fitres \
+	--outDirImgMC=${outDirMC}/img   --outDirImgData=${outDirData}/img  --commonCut $commonCut $NOPUOPT $FLOATTAILSOPT --imgFormat $imgFormat > ${outDirData}/log/validation.log||  exit 1 "
+echo "making table"
+echo "
+./script/makeTable.sh --regionsFile ${regionFile}  --runRangesFile ${runRangesFile} --commonCut=${commonCut} \
+    --outDirFitResMC=${outDirMC}/fitres --outDirFitResData=${outDirData}/fitres \
+    >  ${outDirTable}/monitoring_stability-${invMass_var}-${selection}.tex || exit 1"
 
 ./script/makeTable.sh --regionsFile ${regionFile}  --runRangesFile ${runRangesFile} --commonCut=${commonCut} \
     --outDirFitResMC=${outDirMC}/fitres --outDirFitResData=${outDirData}/fitres \
     >  ${outDirTable}/monitoring_stability-${invMass_var}-${selection}.tex || exit 1
 
+echo "made table"
 
 
 ###################### Make stability plots
@@ -200,12 +238,29 @@ if [ ! -d ${outDirData}/img/stability/$xVar ];then
     mkdir -p ${outDirData}/img/stability/$xVar
 fi
 
+echo "Make stability plots"
+
 ./script/stability.sh -t  ${outDirTable}/monitoring_stability-${invMass_var}-${selection}.tex \
     --outDirImgData ${outDirData}/img/stability/$xVar/ -x $xVar -y peak || exit 1
 ./script/stability.sh -t  ${outDirTable}/monitoring_stability-${invMass_var}-${selection}.tex \
     --outDirImgData ${outDirData}/img/stability/$xVar/ -x $xVar -y scaledWidth || exit 1
 fi
 
+if [ $USER = "lcorpe" ]; then
+#WEBDIR=/afs/cern.ch/user/l/lcorpe/www/MoCa/stability/$baseDir
+WEBDIR=/afs/cern.ch/user/l/lcorpe/www/MoCa/validation/$baseDir
+mkdir -p $WEBDIR/slides
+mkdir -p $WEBDIR/stability
+cp ${outDirData}/img/stability/$xVar/*eps $WEBDIR/stability/.
+cd $WEBDIR/stability/.
+echo $PWD
+for i in *.eps ; do convert "$i" "${i%.*}.pdf" ; done
+for i in *.pdf ; do convert "$i" "${i%.*}.png" ; done
+cp /afs/cern.ch/user/l/lcorpe/public/index.php .
+cd -
+
+echo "[INFO] Stability plots at https://lcorpe.web.cern.ch/lcorpe/MoCa/stability/$baseDir"  
+fi
 
 
 if [ -n "$SLIDES" ];then
@@ -214,12 +269,13 @@ if [ -n "$SLIDES" ];then
     dirData=`dirname $dirData` # remove the selection subdir
     dirMC=`dirname $outDirMC` # remove the invariant mass subdir
     dirMC=`dirname $dirMC` # remove the selection subdir
-    for file in ${dirData}/${selection}/${invMass_var}/img/*.eps;
+    for file in ${dirData}/${selection}/${invMass_var}/img/*.pdf;
     do
 	echo "${file%????}.pdf" #remove four last characters ".eps" by ".pdf"
-	if [ ! -e "${file%????}.pdf" ];then
-	    echo "processing Data $file" && epstopdf $file
-	fi;
+	#if [ ! -e "${file%????}.png" ];then
+	    echo "processing Data $file" && convert "$file" "${file%.*}.png"
+	   # echo "processing Data $file" && convert "$file" "${file%.*}.pdf"
+#	fi;
     done
    
     
@@ -228,8 +284,14 @@ if [ -n "$SLIDES" ];then
 		mkdir $baseDir/slides
 	
 	if [ $USER == "lcorpe" -o $USER == "lbrianza" ];then
+
+#WEBDIR=/afs/cern.ch/user/l/lcorpe/www/MoCa/stability/$baseDir
+WEBDIR=/afs/cern.ch/user/l/lcorpe/www/MoCa/validation/$baseDir
+mkdir -p $WEBDIR/slides
+mkdir -p $WEBDIR/stability
+
 	#for Louie and Luca, lazy: automatically replace variable names and compile latex...
-	cat $configFile | grep s1 >tmp/mc.txt
+	cat $configFile | grep "s1 " | grep -v "#" >tmp/mc.txt
 	cat $configFile | grep d1 >tmp/data.txt
 	sed -e "s|_|-lc|g" tmp/data.txt >tmp/data.1.txt
 	sed -e "s|\/|!|g" tmp/data.1.txt >tmp/data.txt
@@ -237,43 +299,78 @@ if [ -n "$SLIDES" ];then
 	sed -e "s|_|-lc|g" tmp/mc.1.txt >tmp/mc.txt
 	#cp tmp/data.1.txt tmp/data.txt 
 	#cp tmp/mc.1.txt tmp/mc.txt 
-	rerecoTag=`cut -d "!" -f16 tmp/data.txt`
-	runRange=`cut -d "!" -f18 tmp/data.txt`
-	json=`cut -d "!" -f19 tmp/data.txt`
-	dataNtuples=`cut -d ":" -f2 tmp/data.txt`
-	mcNtuples=`cut -d ":" -f2 tmp/mc.txt`
+	#rerecoTag=`cut -d "!" -f16 tmp/data.txt`
+	#runRange=`cut -d "!" -f18 tmp/data.txt`
+	#json=`cut -d "!" -f19 tmp/data.txt`
+	#dataNtuples=`cut -d ":" -f2 tmp/data.txt`
+	#mcNtuples=`cut -d ":" -f2 tmp/mc.txt`
 	echo $CMSSW_VERSION >  tmp/out.txt
-	cmssw_version=`sed "s|_|-lc|g" tmp/out.txt`
+	#cmssw_version=`sed "s|_|-lc|g" tmp/out.txt`
 	echo $invMass_var >  tmp/out.txt
-	invmass_var=`sed "s|_|-lc|g" tmp/out.txt`
+	#invmass_var=`sed "s|_|-lc|g" tmp/out.txt`
 	echo $invmass_var
 	echo "$rerecoTag "
 	echo "$runRange "
 	echo "$json "
 	echo "$dataNtuples "
+	echo "$mcNtuples "
 	echo "$cmssw_version "
-	
+	echo " -e 's/0MCNtuples/$mcNtuples/g' \ "
 
-  sed -e "s/0intLumi/19.636 (\/pb)/g" \
-	-e "s/0runRange/$runRange/g" \
-	-e "s/0globalTag/$GLOBALTAG/g" \
-	-e "s/0rerecoTag/$rerecoTag /g" \
-	-e "s/0gitTag/$GITTAG/g" \
-	-e "s/0cmssw/$cmssw_version/g" \
-	-e "s/0dataNtuples/$dataNtuples/g" \
+#	-e "s/0runRange/$runRange/g" \
+#	-e "s/0globalTag/$GLOBALTAG/g" \
+#	-e "s/0rerecoTag/$rerecoTag /g" \
+#	-e "s/0cmssw/$cmssw_version/g" \
+#	-e "s/0gitTag/$GITTAG/g" \
+#	-e "s/0dataNtuples/$dataNtuples/g" \
+
+  sed -e "s/0intLumi/$intLumi (\/pb)/g" \
 	-e "s/0MCNtuples/$mcNtuples/g" \
 	-e "s/0floatingTail/$FLOATTAILSTEXT/g" \
 	-e "s/std. SC energy/$rerecoTag $invmass_var/g" \
   -e "s/0puReweight/Reweight using nPV distributions in data and MC/g" \
 	-e "s/0json/$json/g" tmp/validation-${invMass_var}-slides.tex > tmp/tmp.tex 
 	
+	
 	sed -e "s|!|\/|g" tmp/tmp.tex >tmp/tmp.1.tex
 	sed -e "s|13TeV|13TeV |g" tmp/tmp.1.tex >tmp/tmp.2.tex
 	sed -e "s|-lc|\\\_|g" tmp/tmp.2.tex > tmp/validation-${invMass_var}-slides.tex
-	#	rm tmp/tmp.tex
-    pdflatex tmp/validation-${invMass_var}-slides.tex
+   	rm tmp/tmp.tex
+    pdflatex -interaction nonstopmode -file-line-error tmp/validation-${invMass_var}-slides.tex
     cp tmp/validation-${invMass_var}-*.tex $baseDir/slides/. 
+    cp tmp/validation-${invMass_var}-*.tex $WEBDIR/slides/. 
     cp validation-${invMass_var}-slides.pdf  $baseDir/slides/validation-${invMass_var}-slides-$rerecoTag.pdf
+		cp $baseDir/slides/validation-${invMass_var}-slides-$rerecoTag.pdf $WEBDIR/slides/.
+    
+		mkdir -p $WEBDIR/$selection/validation-fits/
+		echo "[ INFO ] making $WEBDIR/$selection/validation-fits/ " 
+		cp  ${outDirData}/img/*{png,pdf} $WEBDIR/$selection/validation-fits/.
+		cp ~lcorpe/public/index.php $WEBDIR/$selection/validation-fits/.
+
+		cp ${outDirTable}/monitoring_summary-${invMass_var}-${selection}.tex $WEBDIR/slides/.
+
+		##compare vs reference colelction
+		mkdir $WEBDIR/comparison-vs-ref
+		./script/compareRereco.sh $REF/dato/$REF/${selection}/invMass_rawSC/table/monitoring_summary-invMass_rawSC-${selection}.tex  ${outDirTable}/monitoring_summary-${invMass_var}-${selection}.tex
+
+		cp comparison/comparison.tex $WEBDIR/comparison-vs-ref/comparison_rev_vs_$baseDir.tex
+		cp comparison.pdf $WEBDIR/comparison-vs-ref/comparison_rev_vs_$baseDir.pdf
+		cp $REF/dato/$REF/${selection}/invMass_rawSC/table/monitoring_summary-invMass_rawSC-${selection}.tex $WEBDIR/comparison-vs-ref/reference-table.tex
+		cp ${outDirTable}/monitoring_summary-${invMass_var}-${selection}.tex $WEBDIR/comparison-vs-ref/$baseDir-table.tex
+		cp comparison/slide3.tex $WEBDIR/comparison-vs-ref/comparison_rev_vs_deltaM.tex
+		cp comparison/slide5.tex $WEBDIR/comparison-vs-ref/comparison_rev_vs_deltaP.tex
+		cp comparison/slide6.tex $WEBDIR/comparison-vs-ref/comparison_rev_vs_sigmaCB.tex
+		cp comparison/slide8.tex $WEBDIR/comparison-vs-ref/comparison_rev_vs_resol.tex
+		cp comparison/slide10.tex $WEBDIR/comparison-vs-ref/comparison_rev_vs_addSmearing.tex
+		cp comparison/slide11.tex $WEBDIR/comparison-vs-ref/comparison_rev_vs_chi2.tex
+		cp comparison/slide15.tex $WEBDIR/comparison-vs-ref/comparison_rev_vs_sigmaEff.tex
+		cp comparison/slide2.tex $WEBDIR/comparison-vs-ref/comparison_nevents.tex
+
+		cp $WEBDIR/comparison-vs-ref/*tex $baseDir/slides/.
+
+		echo "$baseDir/slides/validation-${invMass_var}-slides-$rerecoTag.pdf"
+		echo "Slides at $WEBDIR/slides/."
+		echo "on web at:  https://lcorpe.web.cern.ch/lcorpe/MoCa/validation/$baseDir"
 	fi
 		
 fi
